+++ hdl-modules-tester/generator.sv
@@ -1,0 +1,1 @@
+../verilator_tests/generator.sv