+++ verilator_tests/main_udp.h
@@ -1,3 +1,0 @@
-#pragma once
-
-int runUdpEngineMain(int argc, char** argv);